python/requirements.txt
mesa>=3.4.1
typedb-driver==3.8.1
pandas>=2.0
